fix: setup webhook client
