--- src/lib.rs
@@ -1,12 +1,0 @@
-pub mod client;
-pub mod config;
-pub mod context;
-pub mod error;
-pub mod message;
-pub mod queue;
-pub mod token;
-pub mod utils;
-pub mod web;
-
-// Re-exports
-pub use error::{Error, Result};
