+++ webhook-client/Cargo.toml
@@ -24,3 +24,5 @@ tracing-subscriber.workspace = true
 tungstenite.workspace = true
 url.workspace = true
 uuid.workspace = true
+webhook = { path = "../webhook" }
+zerror = { path = "../zerror" }
